billing_api/tests/functional/requirements.txt
pytest==6.2.5
pytest-asyncio==0.15.1
pydantic==1.8.2
httpx==0.19.0
psycopg2-binary==2.8.6
